fix(services): move gemini apis to services folder
Moves gemini apis to a new services folder to improve organization and modularity of the codebase.
